fix deleteFavourite

+++ src/controller/FavouriteController.js
@@ -64,9 +64,9 @@ const getAll = async (userId) => {
     }
 }
 
-const remove = async (tourId) => {
+const remove = async (tourId, userId) => {
     try {
-        await _Favourite.deleteOne({ tourId: tourId });
+        await _Favourite.deleteOne({ tourId: tourId, userId: userId });
         return 0;
     }
     catch (error) {
